requirements.txt
Flask==2.3.2
tensorflow==2.13.0
opencv-python==4.9.0.80
mysqlclient==2.2.0
numpy==1.24.4
Pillow==9.5.0
geopy==2.4.1
requests==2.31.0
urllib3==2.2.1
